finalize admin read function, move some shared utils into util, clean up admin page
